Finished HELP.md and added ability to get log text through GET /log.

+++ src/main/java/com/zackmurry/zmdb/files/FileEditor.java
@@ -384,6 +384,10 @@ public class FileEditor {
         File[] files = file.listFiles();
         if(files != null) {
             for (File subdir : files) {
+                if(subdir.getName().equals("details.txt")) {
+                    setIndexOfTable(subdir, "NULL"); //clearing the index column
+                    continue;
+                }
                 if(deleteFolder(subdir) != 1) return 0;
             }
         }
